# Implement hash function