Duplicates improvement made to Base's Enums: JuliaLang/julia#30500

**Before:**
```julia
julia> @bitflag Flag1 flag1=0 flag2 flag3 flag4

julia> @benchmark hash(flag1)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  11.035 ns … 52.801 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.714 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   16.003 ns ±  1.305 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                        ▁▃ █▁▄        ▁
  ▂▂▂▁▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▃▃▃▃▃▄▂▃▄██▆███▃▃▂▂▂▂▆▅█▆▆█▅▃▂ ▃
  11 ns           Histogram: frequency by time        17.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

**After:**
```julia
julia> @benchmark hash(flag1)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  1.397 ns … 33.455 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.819 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.907 ns ±  1.042 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                               ██    ▅▅
  ▂▁▁▁▁▂▂▂▁▁▁▁▂▃▂▁▁▁▁▂▂▁▁▁▂▃▃▂▁▁▁▂▂▂▁▁▁▁▂▂▂▁▁▃███▆▁▂▄███▃▁▂▂ ▃
  1.4 ns         Histogram: frequency by time        5.66 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```